terraform/common.tf: Enforce string type for some boolean values
This fixes a backward compatibility issue with 0.12 terraform where if
not specified explicitly, the "true" is not considered as boolean.

Without the change, eco_enable_tls = "true" is not the same as
eco_enable_tls = true, thus running the terraform apply with the new eco
module will delete the tls certs.

Alternatively, we should change all the types to boolean, but that will
cause some backward compatibility issue as mentioned above.